+++ tests-data/struct-class/out.gen.js
@@ -1,11 +1,20 @@
 import * as bare from "@bare-ts/lib"
-import * as ext from "./ext.js"
 
+
+export class Person {
+    constructor(
+        name,
+        age,
+    ) {
+        this.name = name
+        this.age = age
+    }
+}
 
 export function readPerson(bc) {
     const name = (bare.readString)(bc)
     const age = (bare.readU8)(bc)
-    return ext.Person(name, age)
+    return new Person(name, age)
 }
 
 export function writePerson(bc, x) {